Add differentiation for Spawn() in encoding
- If analyzeSpawnable flag is turned on during encoding, the encoder
will create copies of encoded sequence for each spawnable agent. Each
copy will have 2 types of Spawn(): the normal Spawn() indicates it is an
action that used to spawn another agent. The @Spawn (used during
encoding only) indicates it is an action used to spawn the corresponding
agent.
- Data duplication (in case of sequence with multiple Spawn() that take
same agent identifier) in DatabaseCreator is removed